Mark y as freevar in x@y
Revert f48f6e05706e2b03bc84307f0dab7e189149edd5.

For records, y is not free in x@y; it is a field name.

For lists, y *is* free in x@y; it is an index expression (could be a var).

For now, we'll assume it might be an expression and mark it as a
(possibly extra) freevar.